[front/lib/resources] - refactor: centralize fetchOutputItemsByActionIds logic
 - Extracted repeated fetchOutputItemsByActionIds logic into a separate static method in AgentMCPActionResource
 - Replaced inline querying of output items in `enrichActionsWithOutputItems` method with a call to the new centralized method

[front/temporal/analytics_queue] - refactor: utilize resources for querying server configurations

 - Replaced direct Model queries with resource-based methods for fetching MCP server configurations
 - Simplified transformation of server configuration IDs to sIDs by leveraging the new resource methods in analytical activities
 - Updated extractRetrievalOutputs to use the new method in AgentMCPActionResource for fetching output items by action IDs
